+++ theme-scripts/common.js
@@ -1,0 +1,6 @@
+export const writeFileCb = (file, err) => {
+  if (err) console.log(err);
+  else {
+    console.log(`${file} updated`);
+  }
+};